refraction

+++ src/main/java/com/example/leavetracker/models/request/LeaveRequestModel.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotNull;
 public class LeaveRequestModel {
 
     @NotNull
-    private Integer EmployeeId;
+    private Long EmployeeId;
 
     @NotNull
     private String leaveStartDate;
